phase2/requirements.txt
dotenv
whisper
torch
sounddevice
coqui-tts
transformers
numpy==2.1
scipy 
gtts 
openai-whisper
requests
groq 
python-dotenv
streamlit 
streamlit-webrtc
elevenlabs
av
streamlit-audio-recorder

